MAINT: Move set and get solution vals methods
